+++ src/main/kotlin/racing/Vehicle.kt
@@ -1,0 +1,5 @@
+package racing
+
+fun interface Vehicle {
+    fun move(count: Int)
+}
